ADDED DEFAULT MODEL CLASS AND ROLE
The default model class is an abstract taking care of basics.
The goal is to have a common API for all models.

Added a Role model to handle registered roles and associated salaries.

Updated Account model to use the Model abstract.
Also added a few properties that could be useful.